+++ typings/zotero.d.ts
@@ -4,7 +4,7 @@
 declare global {
     const Zotero: _ZoteroTypes.Zotero;
     const IOUtils: typeof import("resource://gre/modules/IOUtils.sys.mjs").IOUtils;
-    const Services: typeof import("resource://gre/modules/Services.jsm").Services;
+    const Services: typeof import("resource://gre/modules/Services.sys.mjs").Services;
     const Components: nsIXPCComponents;
     const PathUtils: typeof import("resource://gre/modules/PathUtils.sys.mjs").PathUtils;